added_libft

+++ shapes.c
@@ -6,29 +6,11 @@
 /*   By: scornaz <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/27 12:49:25 by scornaz           #+#    #+#             */
-/*   Updated: 2017/12/28 12:10:53 by scornaz          ###   ########.fr       */
+/*   Updated: 2017/12/28 14:51:09 by scornaz          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
-
-/* void	square(int x, int y, int width, t_libx *libx) */
-/* { */
-/* 	int tmp; */
-/* 	int tmp2; */
-
-/* 	tmp = 0; */
-/* 	while (tmp < width) */
-/* 	{ */
-/* 		tmp2 = 0; */
-/* 		while (tmp2 < width) */
-/* 		{ */
-/* 			PUT(y + tmp2, x + tmp); */
-/* 			++tmp2; */
-/* 		} */
-/* 		++tmp; */
-/* 	} */
-/* } */
 
 void	draw_line(t_line line, int reverse, long color, t_libx *libx)
 {
@@ -37,14 +19,14 @@ void	draw_line(t_line line, int reverse, long color, t_libx *libx)
 
 	y = 0;
 	x = 0;
-	while ((int)x != line.steps)
+	while (x != line.steps)
 	{
 		if (reverse)
 		{
-			PUT(line.origin.x + (int)y, line.origin.y + (int)x, color);
+			PUT(line.origin.x + y, line.origin.y + x, color);
 		}
 		else
-			PUT(line.origin.x + (int)x, line.origin.y + (int)y, color);
+			PUT(line.origin.x + x, line.origin.y + y, color);
 		x += line.steps > 0 ? 1 : -1;
 		y += line.slope;
 	}
